--- src/app/components/add-visualization-panel/add-visualization-panel.component.html
@@ -1,29 +1,0 @@
-<div [class.neon-green-dark-theme]="themesService.getCurrentTheme().id === 'neon-green-dark-theme'">
-    <div class="add-visualization-grid-container" mat-fill>
-        <mat-grid-list cols="4" rowHeight="130px" gutterSize="5px">
-            <mat-grid-tile *ngFor="let item of visualizations; let i = index;" [class.selected]="item.selected"
-                [colspan]="1" [rowspan]="1" class="add-visualization-tile" matRipple>
-                <a (click)="onItemSelected($event.shiftKey, i)">
-                    <div class="neon-popup-thumbnail center-block text-center">
-                        <div class="center-block">
-                            <div class="center-block icon-container">
-                                <img class="neon-svg-icon" src="./assets/icons/{{item.icon}}.svg">
-                            </div>
-                            <div class="center-contents">
-                                <span class="text-uppercase">{{ item.name }}</span>
-                            </div>
-                        </div>
-                    </div>
-                </a>
-            </mat-grid-tile>
-        </mat-grid-list>
-    </div>
-    <mat-divider></mat-divider>
-    <div class="checkmark-container">
-        <input type="checkbox" [checked]="rightPanelService.showVisShortcut" (change)="rightPanelService.showVisShortcut = !rightPanelService.showVisShortcut">
-        Show shortcut
-    </div>
-    <div mat-fill class="flex">
-        <span class="fill-remaining"></span>
-    </div>
-</div>
