--- src/app/components/add-visualization-panel/add-visualization-panel.component.scss
@@ -1,82 +1,0 @@
-/*
- * Copyright 2017 Next Century Corporation
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
- @import '../base-neon-component/base-neon.component.scss';
-
-.add-visualization-grid-container {
-    height: 100%;
-    margin: 10px;
-    overflow-y: auto;
-    /* Add padding between tile and scrollbar. */
-    padding-right: 10px;
-    width: 100%;
-}
-
-.add-visualization-message {
-    display: inline;
-    font-weight: bold;
-    line-height: 15px;
-    margin: 0px;
-    padding: 10px;
-}
-
-.add-visualization-tile {
-    min-height: 50px;
-    min-width: 50px;
-    padding: 0px;
-    display: block;
-    box-sizing: border-box;
-
-    a {
-        background: none;
-        border: 1px solid transparent;
-        display: block;
-        padding: 15px 5px;
-        outline: 0;
-        width: 100%;
-
-        &.collapsed:hover,
-        &:focus,
-        &:hover {
-            cursor: pointer;
-            color: mat-color($primary);
-        }
-    }
-}
-
-.center-contents {
-    justify-content: center;
-    text-align: center;
-    display: flex;
-}
-
-.add-visualization-tile:hover {
-    border: 1px solid mat-color($primary);
-    border-radius: 7px;
-}
-
-.add-visualization-tile:active {
-    border: 1px solid  mat-color($primary);
-    border-radius: 7px;
-}
-
-.app-simple-snack-center {
-    flex: 0 0 auto;
-    font-family: $mat-font-family;
-}
-
-.checkmark-container {
-    padding: 15px;
-}
