--- src/app/components/add-visualization-panel/add-visualization-panel.component.spec.ts
@@ -1,99 +1,0 @@
-/*
- * Copyright 2017 Next Century Corporation
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { By, DomSanitizer } from '@angular/platform-browser';
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, ElementRef, Injector, ViewEncapsulation } from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { MatDialogRef, MatSnackBar } from '@angular/material';
-import {} from 'jasmine-core';
-
-import { AddVisualizationPanelComponent } from './add-visualization-panel.component';
-
-import { ActiveGridService } from '../../services/active-grid.service';
-import { ConnectionService } from '../../services/connection.service';
-import { DatasetService } from '../../services/dataset.service';
-import { ErrorNotificationService } from '../../services/error-notification.service';
-import { ExportService } from '../../services/export.service';
-import { FilterService } from '../../services/filter.service';
-import { RightPanelService } from '../../services/right-panel.service';
-import { ThemesService } from '../../services/themes.service';
-import { VisualizationService } from '../../services/visualization.service';
-
-import { AppMaterialModule } from '../../app.material.module';
-import { DatabaseMetaData, FieldMetaData, TableMetaData } from '../../dataset';
-import { DatasetServiceMock } from '../../../testUtils/MockServices/DatasetServiceMock';
-import { FilterServiceMock } from '../../../testUtils/MockServices/FilterServiceMock';
-import { NeonGTDConfig } from '../../neon-gtd-config';
-import { neonVariables } from '../../neon-namespaces';
-import * as neon from 'neon-framework';
-import { initializeTestBed } from '../../../testUtils/initializeTestBed';
-
-// Must define the test component.
-@Component({
-        selector: 'app-test-add-visualization-panel',
-        templateUrl: 'add-visualization-panel.component.html',
-        styleUrls: ['add-visualization-panel.component.scss'],
-        encapsulation: ViewEncapsulation.Emulated,
-        changeDetection: ChangeDetectionStrategy.OnPush
-})
-
-class TestAddVisualizationPanelComponent extends AddVisualizationPanelComponent {
-    constructor(
-        activeGridService: ActiveGridService,
-        rightPanelService: RightPanelService,
-        themesService: ThemesService,
-        snackBar: MatSnackBar
-    ) {
-
-        super(
-            activeGridService,
-            rightPanelService,
-            snackBar,
-            themesService
-        );
-    }
-
-    // TODO Add any needed custom functions here.
-}
-
-describe('Component: Sample', () => {
-    let component: TestAddVisualizationPanelComponent;
-    let fixture: ComponentFixture<TestAddVisualizationPanelComponent>;
-    let getService = (type: any) => fixture.debugElement.injector.get(type);
-
-    initializeTestBed({
-        declarations: [
-            TestAddVisualizationPanelComponent
-        ],
-        providers: [
-            ActiveGridService,
-            ThemesService
-        ],
-        imports: [
-            AppMaterialModule,
-            BrowserAnimationsModule,
-            FormsModule
-        ]
-    });
-
-    beforeEach(() => {
-        fixture = TestBed.createComponent(TestAddVisualizationPanelComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-});
