+++ src/app/components/add-visualization/add-visualization.component.scss
@@ -16,12 +16,12 @@
  @import '../base-neon-component/base-neon.component.scss';
 
 .add-visualization-grid-container {
-    height: 400px;
+    height: 100%;
     margin: 10px;
     overflow-y: auto;
     /* Add padding between tile and scrollbar. */
     padding-right: 10px;
-    width: 560px;
+    width: 100%;
 }
 
 .add-visualization-message {
@@ -33,8 +33,8 @@
 }
 
 .add-visualization-tile {
-    min-height: 130px;
-    min-width: 120px;
+    min-height: 50px;
+    min-width: 50px;
     padding: 0px;
     display: block;
     box-sizing: border-box;
@@ -75,4 +75,8 @@
 .app-simple-snack-center {
     flex: 0 0 auto;
     font-family: $mat-font-family;
+}
+
+.checkmark-container {
+    padding: 15px;
 }
