+++ src/app/components/right-panel/right-panel.component.html
@@ -8,7 +8,7 @@
             <app-about-neon></app-about-neon>
         </div>
         <div class="add-visualization-container" *ngIf="rightPanelService.showAddVis">
-            <app-add-visualization-panel></app-add-visualization-panel>
+            <app-add-visualization></app-add-visualization>
         </div>
         <div class="setting-container" *ngIf="rightPanelService.showSettings">
             <app-settings></app-settings>
